requirements.txt
apiclient==1.0.4
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
lxml==4.9.0
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
pandas==1.4.2
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
schedule==1.1.0
pytelegrambotapi==4.5.1
